refactored to reduce redundancy and get rid of unnecessary parameters-
