+++ FileShareManager.cs
@@ -2,28 +2,35 @@
 using Microsoft.Azure.Storage.File;
 using NLog.Common;
 using System;
+using System.IO;
+using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace NLog.Extensions.AzureFileSystem
 {
-    public static class FileShareHelper
+    public class FileShareManager
     {
-        public static CloudFileShare InitializeFileShare(IAzureFileSystemTarget target, CloudFileShare fileShare, string folderName, string fileName)
+        private readonly ReaderWriterLockSlim _fileLock = new ReaderWriterLockSlim();
+        private readonly IAzureFileSystemTarget target;
+        private CloudFileShare _fileShare;
+
+        public FileShareManager(IAzureFileSystemTarget target)
         {
-            return AsyncHelper.RunSync(async () => await InitializeFileShareAsync(target, fileShare, folderName, fileName));
+            this.target = target;
         }
 
-        public static async Task<CloudFileShare> InitializeFileShareAsync(IAzureFileSystemTarget target, CloudFileShare fileShare, string folderName, string fileName)
+        private async Task InitializeFileShareAsync(string folderName)
         {
-            if (fileShare == null)
+            if (_fileShare == null)
             {
-                fileShare = await CreateStorageAccountFromConnectionStringAsync(target)
+                _fileShare = await CreateStorageAccountFromConnectionStringAsync(target)
                     .ConfigureAwait(false);
             }
 
             folderName = folderName.Replace(@"\", @"/");
 
-            var rootDir = fileShare.GetRootDirectoryReference();
+            var rootDir = _fileShare.GetRootDirectoryReference();
 
             // Get a reference to the directory.
             var folder = rootDir.GetDirectoryReference(folderName);
@@ -47,19 +54,18 @@ namespace NLog.Extensions.AzureFileSystem
                     }
                 }
             }
-
-            return fileShare;
         }
-
+        
         /// <summary>
         ///     Validates the connection string information in app.config and throws an exception if it looks like
         ///     the user hasn't updated this to valid values.
         /// </summary>
         /// <param name="storageConnectionString">The storage connection string</param>
         /// <returns>CloudStorageAccount object</returns>
-        private static async Task<CloudFileShare> CreateStorageAccountFromConnectionStringAsync(IAzureFileSystemTarget target)
+        private async Task<CloudFileShare> CreateStorageAccountFromConnectionStringAsync(IAzureFileSystemTarget target)
         {
             CloudFileShare fileShare;
+
             try
             {
                 var storageAccount = CloudStorageAccount.Parse(target.StorageConnectionString);
@@ -92,6 +98,56 @@ namespace NLog.Extensions.AzureFileSystem
             }
 
             return fileShare;
+        }
+
+        public void LogMessageToAzureFile(string logMessage, string folderName, string fileName)
+        {
+            AsyncHelper.RunSync(async () => await LogMessageToAzureFileAsync(logMessage, folderName, fileName));
+        }
+        public async Task LogMessageToAzureFileAsync(string logMessage, string folderName, string fileName)
+        {
+            await InitializeFileShareAsync(folderName);
+
+            if (!_fileLock.IsWriteLockHeld)
+            {
+                _fileLock.EnterWriteLock();
+            }
+
+            try
+            {
+                var rootDir = _fileShare.GetRootDirectoryReference();
+                // Get a reference to the directory.
+                var folder = rootDir.GetDirectoryReference(folderName);
+                var sourceFile = folder.GetFileReference(fileName);
+                var sourceFileExists = await sourceFile.ExistsAsync().ConfigureAwait(false);
+                var messageBytes = Encoding.UTF8.GetBytes(logMessage);
+
+                //Ensure that the file exists.
+                if (!sourceFileExists)
+                {
+                    await sourceFile.CreateAsync(messageBytes.Length).ConfigureAwait(false);
+
+                    InternalLogger.Trace($"AzureFileSystemTarget - File {fileName} Created");
+                }
+                else
+                {
+                    await sourceFile.ResizeAsync(sourceFile.Properties.Length + messageBytes.Length)
+                        .ConfigureAwait(false);
+                }
+
+                using (var cloudStream = await sourceFile.OpenWriteAsync(null).ConfigureAwait(false))
+                {
+                    cloudStream.Seek(messageBytes.Length * -1, SeekOrigin.End);
+                    await cloudStream.WriteAsync(messageBytes, 0, messageBytes.Length).ConfigureAwait(false);
+                }
+            }
+            finally
+            {
+                if (_fileLock.IsWriteLockHeld)
+                {
+                    _fileLock.ExitWriteLock();
+                }
+            }
         }
     }
 }
